% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisVar.R
\docType{data}
\name{ClinVar_GRCh38}
\alias{ClinVar_GRCh38}
\title{ClinVar_GRCh38 Dataset}
\format{
A data frame with columns: Chr, Position, ID, Ref, Alt, QUAL, FILTER, INFO, ALLELEID, Gwas_trait, CLNSIG, CLNVC, Variant_type, P_value, Gene
}
\description{
This dataset contains variant information from the ClinVar database.
}
