% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisVar.R
\docType{data}
\name{GAD_GRCh38}
\alias{GAD_GRCh38}
\title{GAD_GRCh38 Dataset}
\format{
A data frame with columns Rsid, Gene, Gwas_trait, Disease_class, PubmedID, Chr, Position, Ref, Alt, P_value, Variant_type
}
\usage{
GAD_GRCh38
}
\description{
This dataset contains GWAS information from the GAD database.
}
\keyword{datasets}
