% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisVar.R
\docType{data}
\name{GWAS_catalog_GRCh38}
\alias{GWAS_catalog_GRCh38}
\title{GWAS_catalog_GRCh38 Dataset}
\format{
A data frame with columns: Rsid, Chr_37, Position_37, Gwas_trait, Gene, Variant_type, P_value, Pubmedid, Chr, Position, Ref, Alt
}
\description{
This dataset contains GWAS information from the GWAS Catalog.
}
