% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisVar.R
\name{DisVar}
\alias{DisVar}
\title{Find diseases from the VCF file}
\usage{
DisVar(file = "file_name.vcf")
}
\arguments{
\item{file}{the name of the VCF file.}
}
\value{
A data frame containing variant information from various databases
}
\description{
This function finds diseases from a VCF file by comparing the variants in the VCF file with those in various databases.
}
