% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisVar.R
\docType{data}
\name{GRASP_GRCh38}
\alias{GRASP_GRCh38}
\title{GRASP_GRCh38 Dataset (Split)}
\format{
A data frame with columns: Rsid, Chr_37, Position_37, P_value, Pubmedid, Gwas_trait, Phenotype_description, Chr, Position, Gene, Ref, Alt, Variant_type
}
\description{
This dataset contains GWAS information from the GRASP database.
It is split into four parts (\code{GRASP_GRCh38_part1} to \code{GRASP_GRCh38_part4}) and merged internally before use.
}
