% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisVar.R
\docType{data}
\name{GRASP_GRCh38}
\alias{GRASP_GRCh38}
\title{GRASP_GRCh38 Dataset}
\format{
A data frame with columns Rsid, Chr_37, Position_37, P_value, Pubmedid, Gwas_trait, Phenotype_escription, Chr, Position, Gene, Ref, Alt, Variant_type
}
\usage{
GRASP_GRCh38
}
\description{
This dataset contains GWAS information from the GRASP database.
}
\keyword{datasets}
